// packages/agent-core/src/types/tool.ts
export type ToolSuccessResponse =
  | {
      [x: string]: unknown;
      content: (
        | {
            type: 'text';
            text: string;
            annotations?:
              | {
                  audience?: ('user' | 'assistant')[] | undefined;
                  priority?: number | undefined;
                  lastModified?: string | undefined;
                }
              | undefined;
            _meta?: Record<string, unknown> | undefined;
          }
        | {
            type: 'image';
            data: string;
            mimeType: string;
            annotations?:
              | {
                  audience?: ('user' | 'assistant')[] | undefined;
                  priority?: number | undefined;
                  lastModified?: string | undefined;
                }
              | undefined;
            _meta?: Record<string, unknown> | undefined;
          }
        | {
            type: 'audio';
            data: string;
            mimeType: string;
            annotations?:
              | {
                  audience?: ('user' | 'assistant')[] | undefined;
                  priority?: number | undefined;
                  lastModified?: string | undefined;
                }
              | undefined;
            _meta?: Record<string, unknown> | undefined;
          }
        | {
            type: 'resource';
            resource:
              | {
                  uri: string;
                  text: string;
                  mimeType?: string | undefined;
                  _meta?: Record<string, unknown> | undefined;
                }
              | {
                  uri: string;
                  blob: string;
                  mimeType?: string | undefined;
                  _meta?: Record<string, unknown> | undefined;
                };
            annotations?:
              | {
                  audience?: ('user' | 'assistant')[] | undefined;
                  priority?: number | undefined;
                  lastModified?: string | undefined;
                }
              | undefined;
            _meta?: Record<string, unknown> | undefined;
          }
        | {
            uri: string;
            name: string;
            type: 'resource_link';
            description?: string | undefined;
            mimeType?: string | undefined;
            annotations?:
              | {
                  audience?: ('user' | 'assistant')[] | undefined;
                  priority?: number | undefined;
                  lastModified?: string | undefined;
                }
              | undefined;
            _meta?: Record<string, unknown> | undefined;
            icons?:
              | {
                  src: string;
                  mimeType?: string | undefined;
                  sizes?: string[] | undefined;
                  theme?: 'light' | 'dark' | undefined;
                }[]
              | undefined;
            title?: string | undefined;
          }
      )[];
      _meta?:
        | {
            [x: string]: unknown;
            progressToken?: string | number | undefined;
            'io.modelcontextprotocol/related-task'?:
              | {
                  taskId: string;
                }
              | undefined;
          }
        | undefined;
      structuredContent?: Record<string, unknown> | undefined;
      isError?: boolean | undefined;
    }
  | {
      [x: string]: unknown;
      toolResult: unknown;
      _meta?:
        | {
            [x: string]: unknown;
            progressToken?: string | number | undefined;
            'io.modelcontextprotocol/related-task'?:
              | {
                  taskId: string;
                }
              | undefined;
          }
        | undefined;
    };
export type ToolErrorResponse = {
  error: {
    type: string;
    message: string;
  };
};

export type ToolResponse = ToolSuccessResponse | ToolErrorResponse;